Genericize(!) app.
